replace long long type w/ int64_t

+++ common/binry.cc
@@ -48,11 +48,11 @@ namespace csl
     }
 
     /* conversions to other types */
-    bool binry::to_integer(long long & v) const
+    bool binry::to_integer(int64_t & v) const
     {
-      if( value_.size() < sizeof(long long) ) { return false; }
+      if( value_.size() < sizeof(int64_t) ) { return false; }
 
-      const long long * p = reinterpret_cast<const long long *>(value_.data());
+      const int64_t * p = reinterpret_cast<const int64_t *>(value_.data());
       v = *p;
 
       return true;
@@ -124,9 +124,9 @@ namespace csl
     }
 
     /* conversions from other types */
-    bool binry::from_integer(long long v)
+    bool binry::from_integer(int64_t v)
     {
-      long long * p = reinterpret_cast<long long *>(value_.allocate(sizeof(long long)));
+      int64_t * p = reinterpret_cast<int64_t *>(value_.allocate(sizeof(int64_t)));
       *p = v;
       return true;
     }
